+++ analyzer/src/test/java/nl/basjes/parse/useragent/TestCalculators.java
@@ -23,6 +23,13 @@ import nl.basjes.parse.useragent.calculate.CalculateDeviceBrand;
 import nl.basjes.parse.useragent.calculate.ConcatNONDuplicatedCalculator;
 import nl.basjes.parse.useragent.calculate.MajorVersionCalculator;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static nl.basjes.parse.useragent.UserAgent.AGENT_NAME;
 import static nl.basjes.parse.useragent.UserAgent.AGENT_NAME_VERSION;
@@ -33,6 +40,8 @@ import static nl.basjes.parse.useragent.UserAgent.DEVICE_BRAND;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TestCalculators {
+
+    private static final Logger LOG = LoggerFactory.getLogger(TestCalculators.class);
 
     @Test
     public void testFieldAgentNameVersionFallback() {
@@ -52,6 +61,54 @@ public class TestCalculators {
         assertEquals("Some_Thing 1.2.3", userAgent.getValue(AGENT_NAME_VERSION));
         assertEquals("1", userAgent.getValue(AGENT_VERSION_MAJOR));
         assertEquals("Some_Thing 1", userAgent.getValue(AGENT_NAME_VERSION_MAJOR));
+    }
+
+    private static final class UrlBrandPair{
+        String url;
+        String brand;
+
+        UrlBrandPair(String url, String brand) {
+            this.url = url;
+            this.brand = brand;
+        }
+
+        @Override
+        public String toString() {
+            return url + " --> " + brand;
+        }
+    }
+
+    public static Iterable<UrlBrandPair> urlsAndBrands() {
+        List<UrlBrandPair> list = new ArrayList<>();
+        list.add(new UrlBrandPair("https://yauaa.basjes.nl",                    "Basjes"));
+        list.add(new UrlBrandPair("https://about.censys.io/",                   "Censys"));
+        list.add(new UrlBrandPair("http://www.google.com/bot.html",             "Google"));
+        list.add(new UrlBrandPair("https://github.com/",                        "Unknown"));
+        list.add(new UrlBrandPair("https://github.com/nielsbasjes/yauaa",       "Nielsbasjes"));
+        list.add(new UrlBrandPair("https://gitlab.com/niels/yauaa",             "Niels"));
+        list.add(new UrlBrandPair("https://github.com/acmesh-official/acme.sh", "Acmesh-Official"));
+
+        // Blacklisted
+        list.add(new UrlBrandPair("http://localhost", "Unknown"));
+        list.add(new UrlBrandPair("http://bit.ly/", "Unknown"));
+        list.add(new UrlBrandPair("https://bit.ly/", "Unknown"));
+
+        // Not a domain
+        list.add(new UrlBrandPair("http://50.22.201.16/Wazzup", "Unknown"));
+
+        list.add(new UrlBrandPair("http://amzn.to/1vsZADi", "Amzn"));
+
+        return list;
+    }
+
+    @ParameterizedTest(name = "Test {index} -> Input: \"{0}\"")
+    @MethodSource("urlsAndBrands")
+    public void checkBrandUrlExtraction(UrlBrandPair pair) {
+        LOG.info("URL: {}", pair);
+        MutableUserAgent userAgent = new MutableUserAgent();
+        userAgent.setForced(UserAgent.AGENT_INFORMATION_URL, pair.url, 1);
+        new CalculateDeviceBrand().calculate(userAgent);
+        assertEquals(pair.brand, userAgent.getValue(DEVICE_BRAND));
     }
 
 }